+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All Notable changes to the library will be documented in this file
 
+## 1.0.1 - 2022-02-22
+
+- Added configuration key to customize the temporary directory used for passes.
+- Fix for the path of the default `ffmpeg-passes*` temporary directory.
+
 ## 1.0.0 - 2022-02-09
 
 Upgraded dependencies, integrated the Alchemy Binary library, and dropped support for anything below PHP 8.0